(feat) Add acceptedLanguage
This is a refactor to get code out from getLangFromRequest() which is
now deprecated in favor of the new siteRouter hook.
